Stop installing things automatically in poetry_install
GitHub's syntax for dealing with if/else variables is lacking, but the
alternative is allowing people to pass arbitrary bash to my action. In
order to avoid that, I'm going to stop installing stuff and leave it to
the end user to decide what install steps and run steps they want at the
consumption end. This will necessitate a change to poetry_export as well.